LayoutTests/media/modern-media-controls/scrubber/scrubber-has-correct-ax-label-expected.txt

Testing for an accessibility text on the scrubber that correctly reflects the changes in media.currentTime.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.getAttribute('aria-valuetext') is "0 seconds"
PASS media.currentTime is 2
PASS input.getAttribute('aria-valuetext') === '2 seconds' became true
PASS successfullyParsed is true

TEST COMPLETE
